Move store package into data package (it will become data-specific)
