+++ data/deduplicator/deduplicator.go
@@ -12,9 +12,9 @@ package deduplicator
 
 import (
 	"github.com/tidepool-org/platform/data"
+	"github.com/tidepool-org/platform/data/store"
 	"github.com/tidepool-org/platform/data/types/base/upload"
 	"github.com/tidepool-org/platform/log"
-	"github.com/tidepool-org/platform/store"
 )
 
 type Deduplicator interface {
