+++ data/deduplicator/root/root.go
@@ -15,9 +15,9 @@ import (
 
 	"github.com/tidepool-org/platform/data/deduplicator"
 	"github.com/tidepool-org/platform/data/deduplicator/truncate"
+	"github.com/tidepool-org/platform/data/store"
 	"github.com/tidepool-org/platform/data/types/base/upload"
 	"github.com/tidepool-org/platform/log"
-	"github.com/tidepool-org/platform/store"
 )
 
 func NewFactory() deduplicator.Factory {
