+++ data/deduplicator/truncate/truncate.go
@@ -16,9 +16,9 @@ import (
 	"github.com/tidepool-org/platform/app"
 	"github.com/tidepool-org/platform/data"
 	"github.com/tidepool-org/platform/data/deduplicator"
+	"github.com/tidepool-org/platform/data/store"
 	"github.com/tidepool-org/platform/data/types/base/upload"
 	"github.com/tidepool-org/platform/log"
-	"github.com/tidepool-org/platform/store"
 )
 
 func NewFactory() deduplicator.Factory {
